Ебаный пиздец

+++ curse-work-2021/Program.cs
@@ -17,7 +17,7 @@ namespace Database
             {
                 Processing.Get(s, "Test", out bool onLastLetter, out bool cityIsUsed, out string outCity, out char nextLetter, out int letterNumberFromEnd,
                     out string wikiUrl, out string yandexUrl, out string googleUrl, out string mapUrl,
-                    out (double latitude, double longitude) coordinateCity, out string photoUrl);
+                    out (decimal latitude, decimal longitude) coordinateCity, out string photoUrl);
                 Console.WriteLine(
                     $"onLastLetter {onLastLetter}, cityIsUsed {cityIsUsed}, outCity {outCity}, nextLetter {nextLetter}, letterNumberFromEnd {letterNumberFromEnd}, wikiUrl {wikiUrl}, " +
                     $"yandexUrl {yandexUrl}, googleUrl {googleUrl}, mapUrl {mapUrl}, coordinateCity.latitude {coordinateCity.latitude}, " +
